bpf, doc: fix formatting for emphasis on a directive.
Formatting for the edited lines would make it a citation, which does not
really make sense in this case. Since we want to emphasize the
directive, why not make it bold instead?

Do not touch other citations in the document, that are used for
highlighting emails and links and at least have a consistent look.

(I believe that bold italics does not work with RST? So replace *not*
with NOT in the middle of the sentence.)